--- a/bids/layout/layout.py
+++ b/bids/layout/layout.py
@@ -149,7 +149,7 @@ class BIDSLayout(object):
                 database_path, reset_database, config, init_args)
 
             if indexer is None:
-                indexer = BIDSLayoutIndexer(valid_only=validate, **indexer_kwargs)
+                indexer = BIDSLayoutIndexer(validate=validate, **indexer_kwargs)
             indexer(self)
 
         # Add derivatives if any are found