## The problem
As indicated in #935, the pattern matching was pretty brittle, and did not take well sophisticated patterns.

## The solution
This PR converts string matching patterns into regexes (so only one type of pattern is tested, removed that if/else branch of the validation).

## Side effects
This PR makes several decisions that could be controversial:

* When a ``force_index`` or ``ignore``  pattern matches a folder, that's final. This is useful mostly for exclusions: marking a folder for ignore will not try to index any of its contents despite maybe having nested ``force_index``. The speedup in indexing of this decision is huge. If one still wants to have nested force/ignore, please check the new test cases in ``test_layout.py`` showing how to avoid a regex to exactly match the path of a folder (e.g., instead of `ignore=["models"]` you probably want to `ignore=[re.compile(r"models/.+")]` so that `/models/` is not entered for indexing.
* Revises the default patterns
* Deprecates the `validate` argument of `BIDSLayoutIndexer` in favor of `valid_only` as the dataset is not being validated, but rather you want to specify whether only BIDS-valid files should be returned.
* Accommodates the indexing of derivatives, forcing the generation of a new indexer every time.
* A side effect of the above is that fully invalid datasets will not be picked up as derivatives automatically. TBH, I don't know whether this was a feature or a bug.

## Other
* This PR is working with the regex example of #935 and dramatically speeds up the indexing by MRIQC.
* Happy to update whatever is not well received

Resolves: #935.