@oesteban 2023-02-03
Okay, tests are passing locally. There seems to be an installation problem - but I'm guessing that's unrelated.

I've updated the PR message to cover better what's done here.