@oesteban 2023-02-03
Okay, I can see the pre-release tests also break on master. numpy seems to be the culprit.

@effigies @adelavega - could you have a look into this PR?