@adelavega 2023-02-03
I haven't had time to do a deep dive, but my general comment is that:

- At first glance, I agree with your changes to the policy (specifically `ignore` and `force_index` being final) and the the existing pattern matching was brittle
- My main concern is we should think about upcoming major changes to pybids before making any changes to the current implementation. I rather settle on a final policy for how `force_index` and `ignore` behave (or if `force_index` should be dropped altogether).