@effigies 2023-02-03
@oesteban Without a benchmark dataset where I can verify time savings, it's going to be tough to submit an alternative. I just tried pulling a dataset with a large number of derivatives, and it took 33s to load on master and your branch.