@oesteban 2023-02-03
Rename "derivatives" to "sourcedata".

People have all sorts of nonstandard stuff.

In addition to that, this PR makes pattern matching more reliable (for
instance see the corresponding issue)



On Fri, Feb 3, 2023, 21:31 Chris Markiewicz ***@***.***>
wrote:

> @oesteban <https://github.com/oesteban> Without a benchmark dataset where
> I can verify time savings, it's going to be tough to submit an alternative.
> I just tried pulling a dataset with a large number of derivatives, and it
> took 33s to load on master and your branch.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/bids-standard/pybids/pull/936#issuecomment-1416370829>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAESDRU67QOAWIPNFWMRZC3WVVTLTANCNFSM6AAAAAAUBTLFJA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>