@effigies 2023-02-03
Using

```Python
            ignore_paths = [
                re.compile(r"^(?!/sub-[a-zA-Z0-9]+)"),
                # Exclude modalities and contrasts ignored by MRIQC (doesn't know how to QC)
                re.compile(
                    r"sub-[a-zA-Z0-9]+(/ses-[a-zA-Z0-9]+)?/(dwi|fmap|perf)/"
                ),
                # negative lookbehind to only index T1w, T2w and bold (please note length must be constant)
                re.compile(r"^.+(?<!(_T1w|_T2w|bold))\.(json|nii|nii\.gz)$"),
            ]
```

I go from 6.5s load on `master` to 9s load with this PR.

I'll go ahead and try my proposal and see if I at least don't degrade from here. I'll leave it to you to test on your use case...