@oesteban 2023-02-03
> I go from 6.5s load on master to 9s load with this PR

I assume that's because master is not indexing the sub-* directories.

In any case, with derivatives/ you won't see any difference because of its
special treatment.

I'll benchmark on my dataset and post the tree view of it.

On Fri, Feb 3, 2023, 21:43 Chris Markiewicz ***@***.***>
wrote:

> Using
>
>             ignore_paths = [
>                 re.compile(r"^(?!/sub-[a-zA-Z0-9]+)"),
>                 # Exclude modalities and contrasts ignored by MRIQC (doesn't know how to QC)
>                 re.compile(
>                     r"sub-[a-zA-Z0-9]+(/ses-[a-zA-Z0-9]+)?/(dwi|fmap|perf)/"
>                 ),
>                 # negative lookbehind to only index T1w, T2w and bold (please note length must be constant)
>                 re.compile(r"^.+(?<!(_T1w|_T2w|bold))\.(json|nii|nii\.gz)$"),
>             ]
>
> I go from 6.5s load on master to 9s load with this PR.
>
> I'll go ahead and try my proposal and see if I at least don't degrade from
> here. I'll leave it to you to test on your use case...
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/bids-standard/pybids/pull/936#issuecomment-1416383483>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAESDRQNYLK4FLR4PZLZMI3WVVUXBANCNFSM6AAAAAAUBTLFJA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>