@oesteban 2023-02-03
Another use case I can think is humongous datasets ala UKB, where allowing a regex to pick 3 of 3k subjects is necessary.

This PR  allows the indexer to completely ignore the undesired subjects without looking into their folders. Master looks into all those folders because of that `default_value`

Unfortunately, I cannot benchmark now