@oesteban 2023-02-04
Benchmarks:

This PR:

```Python
In [1]: import re
   ...: from bids.layout import BIDSLayout, BIDSLayoutIndexer

In [2]: ignore_paths = [
   ...:     re.compile(r"^(?!/sub-[a-zA-Z0-9]+)"),
   ...:     re.compile(r"sub-[a-zA-Z0-9]+(/ses-[a-zA-Z0-9]+)?/(dwi|fmap|perf)/"),
   ...:     re.compile(r"^.+(?<!(_T1w|_T2w|bold))\.(json|nii|nii\.gz)$"),
   ...: ]

In [3]: %timeit BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(valid_only=False, ignore=ignore_paths))
194 ms ± 1.52 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: %timeit BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(valid_only=True, ignore=ignore_paths))
213 ms ± 2.54 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: layout = BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(valid_only=False, ignore=ignore_paths))

In [6]: len(layout.get())
Out[6]: 39

In [7]: layout = BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(valid_only=True, ignore=ignore_paths))

In [8]: len(layout.get())
Out[8]: 39
```

Current master (left `valid_only` as a canary to be sure I'm not in the wrong branch):

```Python
In [1]: import re
   ...: from bids.layout import BIDSLayout, BIDSLayoutIndexer

In [2]: ignore_paths = [
   ...:     re.compile(r"^(?!/sub-[a-zA-Z0-9]+)"),
   ...:     re.compile(r"sub-[a-zA-Z0-9]+(/ses-[a-zA-Z0-9]+)?/(dwi|fmap|perf)/"),
   ...:     re.compile(r"^.+(?<!(_T1w|_T2w|bold))\.(json|nii|nii\.gz)$"),
   ...: ]

In [3]: %timeit BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(valid_only=False, ignore=ignore_paths))  # Try to show

ValueError: 'valid_only' is not a recognized entity. If you're sure you want to impose this constraint, set invalid_filters='allow'.

In [4]: %timeit BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(validate=False, ignore=ignore_paths))
1.51 s ± 8.24 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: %timeit BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(validate=True, ignore=ignore_paths))
1.49 s ± 4.94 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [6]: layout = BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(validate=False, ignore=ignore_paths))

In [7]: len(layout.get())
Out[7]: 0
```

The dataset is not that big:

```Text
$ tree -L 3
.
├── CHANGES
├── README
├── dataset_description.json
├── participants.json
├── participants.tsv
├── rawdata
│   └── sub-2022_11_07_15_37_06_STD_1_3_12_2_1107_5_99_3
│       ├── ses-20221107153815
│       ├── ses-20221221171715
│       ├── ses-20230109172617
│       ├── ses-20230116131825
│       └── ses-20230118160434
├── scans.json
├── sub-pilot
│   ├── ses-1
│   │   ├── anat
│   │   ├── dwi
│   │   ├── fmap
│   │   └── sub-pilot_ses-1_scans.tsv
│   ├── ses-2
│   │   ├── anat
│   │   ├── dwi
│   │   ├── fmap
│   │   ├── func
│   │   └── sub-pilot_ses-2_scans.tsv
│   ├── ses-3
│   │   ├── anat
│   │   └── sub-pilot_ses-3_scans.tsv
│   └── ses-4
│       ├── anat
│       ├── dwi
│       ├── fmap
│       ├── func
│       └── sub-pilot_ses-4_scans.tsv
├── task-control_bold.json
└── task-rest_bold.json

24 directories, 12 files
```

The folder `rawdata/` contains four sessions worth of DICOM files. Those are many, but nothing compared to any regular dataset where the dicoms are sitting around (or a datalad, structure, because current exclusion of dotfolders just doesn't work).

I fear the reason you were getting comparable indexing times is because master is actually indexing zero files with the new patterns. This is what happens in default settings (before this benchmark I renamed `rawdata/` -> `sourcedata/`).

PR:
```Python
In [1]: from bids.layout import BIDSLayout, BIDSLayoutIndexer

In [2]: %timeit BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(valid_only=False))
2.54 s ± 23.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

master:
```Python
In [1]: from bids.layout import BIDSLayout, BIDSLayoutIndexer

# Cell 2 is the valid_only canary

In [3]: %timeit BIDSLayout("/data/datasets/hcph-pilot", indexer=BIDSLayoutIndexer(validate=False))
5.04 s ± 35.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Similar timings happen rolling back `sourcedata/` -> `rawdata/` and adding `ignore=["rawdata"]`.